Add styling to new marker form, code re-organize
